FIX: object is now distance from player + extra release button
if you press and hold, you can release the object being held by just releasing the left mouse button. If you just click it, it doesnt release. As a temp solution, ive allowed right click to release also so now both the buttons are in use. I always imagined left mouse button beign a select and right being a back sort of button anyways since its kinda standard so not out of the ordinary. Altho i dont want this to be the perminant solution lolll